+++ src/test/java/ru/zhenyria/telegram_consulting_plugin/test_util/ChatMemberRepositoryTestUtil.java
@@ -3,7 +3,6 @@ package ru.zhenyria.telegram_consulting_plugin.test_util;
 import lombok.experimental.UtilityClass;
 import ru.zhenyria.telegram_consulting_plugin.model.ChatMember;
 
-import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -11,8 +10,8 @@ import java.util.List;
  */
 @UtilityClass
 public class ChatMemberRepositoryTestUtil {
-    private static final ChatMember CHAT_MEMBER_1 = new ChatMember(BigInteger.valueOf(1L), "default_user_1");
-    private static final ChatMember CHAT_MEMBER_2 = new ChatMember(BigInteger.valueOf(2L), "default_user_2");
+    private static final ChatMember CHAT_MEMBER_1 = new ChatMember(1L, "default_user_1");
+    private static final ChatMember CHAT_MEMBER_2 = new ChatMember(2L, "default_user_2");
 
     private static final List<ChatMember> CHAT_MEMBERS = List.of(CHAT_MEMBER_1, CHAT_MEMBER_2);
